# laravel/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API PGE-QROO",
        "description": "Documentación de la API para la Plataforma de Gestión Energética de Quintana Roo. Permite la gestión de dependencias, edificios, presupuestos y análisis de consumo.",
        "version": "1.0.0"
    },
    "paths": {
        "/api/login": {
            "post": {
                "tags": [
                    "Autenticación"
                ],
                "summary": "Iniciar sesión",
                "description": "Obtiene un token de acceso para el sistema.",
                "operationId": "222b2625838e3c57e017987f73598fda",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "contrasena"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "admin@pgeqroo.com"
                                    },
                                    "contrasena": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "password123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Login exitoso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "login exitoso"
                                        },
                                        "access_token": {
                                            "type": "string",
                                            "example": "1|AbCdEf123..."
                                        },
                                        "token_type": {
                                            "type": "string",
                                            "example": "Bearer"
                                        },
                                        "user": {
                                            "$ref": "#/components/schemas/UserAuthResponse"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Credenciales incorrectas"
                    },
                    "422": {
                        "description": "Datos inválidos"
                    }
                }
            }
        },
        "/api/me": {
            "get": {
                "tags": [
                    "Autenticación"
                ],
                "summary": "Obtener usuario autenticado",
                "description": "Devuelve los datos del usuario logueado. Requiere Bearer Token.",
                "operationId": "8367feff40ba10e43187792cad3132c1",
                "responses": {
                    "200": {
                        "description": "Datos del usuario recuperados",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserFullResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autenticado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/logout": {
            "post": {
                "tags": [
                    "Autenticación"
                ],
                "summary": "Cerrar sesión",
                "description": "Elimina el token de acceso actual del usuario. Requiere Bearer Token.",
                "operationId": "ad65cbbd4e9f201619eae184a5961a98",
                "responses": {
                    "200": {
                        "description": "Sesión cerrada correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Sesión cerrada correctamente"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/analisis/ranking-dependencias": {
            "get": {
                "tags": [
                    "Análisis"
                ],
                "summary": "Ranking de consumo por dependencia",
                "description": "Obtiene el ranking de consumo energético agrupado por dependencia. Útil para gráficas comparativas. Permite filtrar por año y mes.",
                "operationId": "e739b97d6c5a410f5116249df058ee75",
                "parameters": [
                    {
                        "name": "año",
                        "in": "query",
                        "description": "Año del consumo (por defecto el año actual)",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 2025
                        }
                    },
                    {
                        "name": "mes",
                        "in": "query",
                        "description": "Mes del consumo (1–12). Si no se envía, se consideran todos los meses",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "maximum": 12,
                            "minimum": 1,
                            "example": 3
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ranking de consumo obtenido correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "periodo": {
                                            "properties": {
                                                "año": {
                                                    "type": "integer",
                                                    "example": 2025
                                                },
                                                "mes": {
                                                    "type": "string",
                                                    "example": "Todos"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "ranking": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/RankingConsumoDependencia"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autenticado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/consumos/carga-masiva": {
            "post": {
                "tags": [
                    "Consumos"
                ],
                "summary": "Carga masiva de consumos energéticos",
                "description": "Permite importar consumos desde un archivo CSV, XLSX, XLS o TXT. Requiere permiso cargar-consumos.",
                "operationId": "66884a3bcb042ab62e3d8abf2572750c",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "archivo"
                                ],
                                "properties": {
                                    "archivo": {
                                        "description": "Archivo de consumos (csv, xlsx, xls, txt)",
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Datos importados con éxito",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Datos importados con exito"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Errores de validación en el archivo",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "El archivo tiene errores de formato"
                                        },
                                        "details": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/ErrorCargaConsumo"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "No autorizado"
                    },
                    "500": {
                        "description": "Error interno al procesar el archivo"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "UserAuthResponse": {
                "title": "Usuario Autenticado (Respuesta Login)",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 5
                    },
                    "nombre": {
                        "type": "string",
                        "example": "Juan Pérez"
                    },
                    "nombre_usuario": {
                        "type": "string",
                        "example": "jperez"
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "juan.perez@ejemplo.com"
                    }
                },
                "type": "object"
            },
            "RankingConsumoDependencia": {
                "title": "Ranking de Consumo por Dependencia",
                "description": "Datos agregados de consumo energético por dependencia",
                "properties": {
                    "id_dependencia": {
                        "type": "integer",
                        "example": 1
                    },
                    "nombre_dependencia": {
                        "type": "string",
                        "example": "Secretaría de Educación"
                    },
                    "total_consumo": {
                        "type": "number",
                        "format": "float",
                        "example": 15432.75
                    },
                    "total_costo": {
                        "type": "number",
                        "format": "float",
                        "example": 32145.5
                    }
                },
                "type": "object"
            },
            "ErrorCargaConsumo": {
                "title": "Error en Carga Masiva de Consumos",
                "description": "Detalle de error detectado en una fila del archivo importado",
                "properties": {
                    "fila": {
                        "type": "integer",
                        "example": 12
                    },
                    "columna": {
                        "type": "string",
                        "example": "consumo_kwh"
                    },
                    "error": {
                        "type": "array",
                        "items": {
                            "type": "string",
                            "example": "El consumo debe ser numérico"
                        }
                    },
                    "valor_erroneo": {
                        "type": "string",
                        "example": "ABC123"
                    }
                },
                "type": "object"
            },
            "UserFullResponse": {
                "title": "Usuario Completo (Endpoint /me)",
                "allOf": [
                    {
                        "$ref": "#/components/schemas/UserAuthResponse"
                    },
                    {
                        "properties": {
                            "activo": {
                                "type": "boolean",
                                "example": true
                            },
                            "ultimo_login": {
                                "type": "string",
                                "format": "date-time",
                                "example": "2025-12-13T05:00:00.000000Z"
                            },
                            "roles": {
                                "type": "array",
                                "items": {
                                    "type": "object",
                                    "example": {
                                        "id": 1,
                                        "name": "Administrador"
                                    }
                                }
                            },
                            "dependencias": {
                                "type": "array",
                                "items": {
                                    "type": "object",
                                    "example": {
                                        "id": 1,
                                        "name": "Secretaría A"
                                    }
                                }
                            }
                        },
                        "type": "object"
                    }
                ]
            },
...
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
...
}